+++ client/python/tests/e2e/cases/__init__.py
@@ -10,6 +10,7 @@ from .test_polygon import *     # noqa: F401, F403
 from .test_query import *       # noqa: F401, F403
 from .test_rect import *        # noqa: F401, F403
 from .test_shapes import *      # noqa: F401, F403
+from .test_shapes3d import *    # noqa: F401, F403
 from .test_shared import *      # noqa: F401, F403
 from .test_system import *      # noqa: F401, F403
 from .test_text import *        # noqa: F401, F403